[TASK] Reduce code duplication in the repositories
Fixes #69
