+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Reduce code duplication in the repositories (#144)
 - Bump the minimal 10.4 Extbase requirement (#137)
 
 ## 4.2.0
