+++ Tests/Functional/Domain/Repository/FrontendUserRepositoryTest.php
@@ -15,8 +15,8 @@ use TYPO3\CMS\Extbase\Persistence\ObjectStorage;
 use TYPO3\TestingFramework\Core\Functional\FunctionalTestCase;
 
 /**
+ * @covers \OliverKlee\FeUserExtraFields\Domain\Repository\AbstractFrontendUserRepository
  * @covers \OliverKlee\FeUserExtraFields\Domain\Repository\FrontendUserRepository
- * @covers \OliverKlee\FeUserExtraFields\Domain\Repository\FrontendUserRepositoryTrait
  * @covers \OliverKlee\FeUserExtraFields\Domain\Repository\DirectPersistTrait
  */
 final class FrontendUserRepositoryTest extends FunctionalTestCase
